Create ehseb-live.html
Ehseb V.1 webpage launched for use on 09/04/2024.

The page calculates the score of the 4 player card game Hokm Thabt / حكم ثابت up to a score of 540.

No need to assign a human calculator anymore, no more scratching heads to figure out the sum of -41 + 23, and no one will be blamed for forgetting the notebook and pen to record the scores. Simply enter the scores into the card shown on the page and let it do the job for you, while maintaining your records in every round.